+++ crates/merman-render/src/svg/parity/class/mod.rs
@@ -21,11 +21,15 @@ use defs::{class_marker_name, class_markers};
 
 mod edge;
 
+mod interface;
+
 mod label;
 use label::{
     class_apply_inline_styles, class_html_div_style, class_html_label_max_width_px,
     class_html_label_metrics, class_html_title_metrics, render_class_html_label,
 };
+
+mod node;
 
 mod note;